+++ src/models/conditions/RelatedToCategory.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Ryssbowh\CraftTriggers\models\conditions;
+
+use craft\elements\Category;
+
+class RelatedToCategory extends RelatedToEntry
+{
+    /**
+     * @inheritDoc
+     */
+    public function getElementType(): string
+    {
+        return Category::class;
+    }
+}