+++ src/models/triggers/CategorySaved.php
@@ -4,9 +4,9 @@ namespace Ryssbowh\CraftTriggers\models\triggers;
 
 use Ryssbowh\CraftTriggers\Triggers;
 use Ryssbowh\CraftTriggers\models\Trigger;
-use craft\base\Element;
 use craft\elements\Category;
-use craft\events\ModelEvent;
+use craft\events\ElementEvent;
+use craft\services\Elements;
 use yii\base\Event;
 
 class CategorySaved extends Trigger
@@ -33,9 +33,12 @@ class CategorySaved extends Trigger
     public function initialize()
     {
         $_this = $this;
-        Event::on(Category::class, Element::EVENT_AFTER_SAVE, function (ModelEvent $e) use ($_this) {
+        Event::on(Elements::class, Elements::EVENT_AFTER_SAVE_ELEMENT, function (ElementEvent $e) use ($_this) {
+            if (!$e->element instanceof Category) {
+                return;
+            }
             Triggers::$plugin->triggers->onTriggerTriggered($_this, [
-                'category' => $e->sender,
+                'category' => $e->element,
                 'isNew' => $e->isNew,
                 'event' => $e
             ]);