+++ src/models/triggers/UserSaved.php
@@ -4,9 +4,8 @@ namespace Ryssbowh\CraftTriggers\models\triggers;
 
 use Ryssbowh\CraftTriggers\Triggers;
 use Ryssbowh\CraftTriggers\models\Trigger;
-use craft\base\Element;
 use craft\elements\User;
-use craft\events\ModelEvent;
+use craft\events\ElementEvent;
 use yii\base\Event;
 
 class UserSaved extends Trigger
@@ -33,9 +32,12 @@ class UserSaved extends Trigger
     public function initialize()
     {
         $_this = $this;
-        Event::on(User::class, Element::EVENT_AFTER_SAVE, function (ModelEvent $e) use ($_this) {
+        Event::on(Elements::class, Elements::EVENT_AFTER_SAVE_ELEMENT, function (ElementEvent $e) use ($_this) {
+            if (!$e->element instanceof User) {
+                return;
+            }
             Triggers::$plugin->triggers->onTriggerTriggered($_this, [
-                'user' => $e->sender,
+                'entry' => $e->element,
                 'isNew' => $e->isNew,
                 'event' => $e
             ]);